consolidate code to build bus list
