+++ server/jsonHandler.js
@@ -12,7 +12,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.writeWhitelist = exports.writeBusList = exports.readBusStatus = exports.readBusList = exports.readWeather = exports.readWhitelist = exports.writeWeather = exports.writeBuses = exports.readData = void 0;
+exports.writeWhitelist = exports.writeBusList = exports.readBusStatus = exports.readBusList = exports.readWeather = exports.readWhitelist = exports.writeWeather = exports.writeBuses = exports.readData = exports.getBuses = void 0;
 const path_1 = __importDefault(require("path"));
 const fs_1 = __importDefault(require("fs"));
 const busesDatafile = path_1.default.resolve(__dirname, "../data/buses.json");
@@ -23,17 +23,39 @@ const busListDatafile = path_1.default.resolve(__dirname, "../data/busList.json"
 const Announcement = require("./model/announcement");
 const Bus = require("./model/bus");
 const Weather = require("./model/weather");
+function getBuses() {
+    return __awaiter(this, void 0, void 0, function* () {
+        // get all the buses and create a list of objects like the following {number:,change:,time:,status:}
+        const buses = yield Bus.find({});
+        const busList = [];
+        buses.forEach((bus) => {
+            busList.push({ number: bus.busNumber, change: bus.busChange, time: bus.time, status: bus.status });
+        });
+        // if change is 0, make it an empty string
+        busList.forEach((bus) => {
+            if (bus.change === 0)
+                bus.change = "";
+            if (bus.time == undefined)
+                bus.time = new Date();
+            if (bus.status === "normal")
+                bus.status = "";
+            // bus.time = bus.time.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: true });
+            if (bus.status === "")
+                bus.time = "";
+        });
+        // sort the list by bus number
+        busList.sort((a, b) => {
+            return a.number - b.number;
+        });
+        return busList;
+    });
+}
+exports.getBuses = getBuses;
 // Load data file. If no file exists creates one
 function readData() {
     return __awaiter(this, void 0, void 0, function* () {
         const weather = yield Weather.findOne({});
-        let buses = yield Bus.find({});
-        buses = buses.map((bus) => ({
-            number: bus.busNumber || '',
-            change: bus.busChange || '',
-            time: bus.time || '',
-            status: bus.status || ''
-        }));
+        let buses = yield getBuses();
         return { buses: buses, weather: weather, announcement: (yield Announcement.findOne({})).announcement };
     });
 }
