+++ server/jsonHandler.ts
@@ -13,20 +13,37 @@ const Weather = require("./model/weather");
 export type BusData = {number: string, change: string | undefined, time: string | undefined, status: string | undefined};
 export type adminData = {address: string};
 
+export async function getBuses() {
+    // get all the buses and create a list of objects like the following {number:,change:,time:,status:}
+    const buses: any[] = await Bus.find({});
+    const busList: any[] = [];
+    buses.forEach((bus: any) => {
+        busList.push({number: bus.busNumber, change: bus.busChange, time: bus.time, status: bus.status});
+    });
+    // if change is 0, make it an empty string
+    busList.forEach((bus: any) => {
+        if (bus.change === 0) bus.change = "";
+        if(bus.time == undefined) bus.time = new Date();
+        if (bus.status === "normal") bus.status = "";
+        // bus.time = bus.time.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: true });
+        if(bus.status === "") bus.time = "";
+    });
+
+    // sort the list by bus number
+    busList.sort((a: any, b: any) => {
+        return a.number - b.number;
+    });
+
+    return busList;
+}
+
 
 // Load data file. If no file exists creates one
 export async function readData() {
 
     
     const weather = await Weather.findOne({})
-    let buses = await Bus.find({});
-
-    buses = buses.map((bus) => ({
-        number: bus.busNumber || '',
-        change: bus.busChange || '',
-        time: bus.time || '', // .toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: true }) || '',
-        status: bus.status || ''
-    }));
+    let buses = await getBuses();
 
     return {buses: buses, weather: weather, announcement: (await Announcement.findOne({})).announcement};
 }
